// src/types/adminPanel.ts
export interface PackagesData {
    id: number;
    name: string;
    duration: string;
    price: string;
    renewalFee: string;
    validityPeriod: string;
    videoCount: string;
    imageCount: string;
    notificationCount: string;
    audioFileLimit: string;
    barcodeEnabled: boolean;
    templateSelectionEnabled: boolean;
    condolenceMessageEnabled: boolean;
    visitorContentSubmissionEnabled: boolean;
    locationAndNavigationEnabled: boolean;
    sharingEnabled: boolean;
    file360DegreeEnabled: boolean;
    displayEnabled: boolean;
    updateCapabilityEnabled: boolean;
    edit: any;
    delete: any;
  }